requirements.txt
astroid==2.2.5
Django==2.2.1
django-mathfilters==0.4.0
django-summernote==0.8.11.4
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.0.0
psycopg2-binary==2.8.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.1
